perf: 主机与主机关系事件处理优化 #1145
主机关系事件监听代码重构。
